+++ ntex/src/web/test.rs
@@ -704,6 +704,9 @@ where
 
     let (system, server, addr) = rx.recv().unwrap();
 
+    // wait for server
+    thread::sleep(std::time::Duration::from_millis(50));
+
     let client = {
         let connector = {
             #[cfg(feature = "openssl")]